Pipeline build & loading
